+++ src/components/games/hangman/index.ts
@@ -1,0 +1,3 @@
+export { HangmanFigure } from "./HangmanFigure";
+export { HangmanKeyboard } from "./HangmanKeyboard";
+export { HangmanGame } from "./HangmanGame";